Add payment integration to ticketing system
A 'Currency' property has been added to the 'TicketTemplate' entity and payment service integrations have been added in 'CreateTicket' and 'GetTicket' Handlers. This change allows the system to handle payments when a ticket is being created. The payment service API has been added to the application configuration and a new domain service 'IPaymentDomainService' has been added to handle payment infrastructure.