Add user documentation and update authorization policies
Added detailed user documentation for different parts of the app. These sections aim to guide the user on how to operate and navigate the application. Additionally, modified authorization policies in the backend to better handle permissions. Updated the convention authorization requirement in particular by adding a new policy `ReadOneConvention`. The UI for managing organizers in main menu was also removed.